+++ tests/testthat/test-parse_deparse.R
@@ -8,7 +8,7 @@ test_that("timestamp string parsing to offsets and deparsing of offsets to times
       offsets <- -1000:1000
       def <- paste(CFt$units$name[u], "since 1992-08-20")
       t <- CFtime(def, c, offsets)
-      time <- t$cal$offsets2time(t$offsets)
+      time <- t$calendar$offsets2time(t$offsets)
       ts <- as_timestamp(t, "timestamp")
       cf2 <- CFtime(def, c)
       tp <- parse_timestamps(cf2, ts)
@@ -18,10 +18,10 @@ test_that("timestamp string parsing to offsets and deparsing of offsets to times
 
   # Add no offsets, test return value
   t <- CFTime$new("days since 1992-08-20", "standard")
-  res <- t$cal$offsets2time()
+  res <- t$calendar$offsets2time()
   expect_equal(nrow(res), 0L)
   t <- CFTime$new("days since 1992-08-20", "utc")
-  res <- t$cal$offsets2time()
+  res <- t$calendar$offsets2time()
   expect_equal(nrow(res), 0L)
 })
 
@@ -33,7 +33,7 @@ test_that("testing calendars with leap years", {
     for (d in c("1996-01-15", "1996-02-29", "1996-04-01")) {
       def <- paste("days since", d)
       t <- CFtime(def, c, c(1:2500, 36501:39000))
-      time <- t$cal$offsets2time(t$offsets)
+      time <- t$calendar$offsets2time(t$offsets)
       ts <- as_timestamp(t, "timestamp")
       cf2 <- CFtime(def, c)
       tp <- parse_timestamps(cf2, ts)
@@ -52,7 +52,7 @@ test_that("testing calendars with leap years", {
 #       offsets <- runif(10000, max = 10000)
 #       def <- paste(CFt$units$name[u], "since 1978-08-20 07:34:12.2")
 #       expect_warning(t <- CFtime(def, c, offsets)) # not ordered
-#       time <- t$cal$offsets2time(t$offsets)
+#       time <- t$calendar$offsets2time(t$offsets)
 #       ts <- as_timestamp(t, "timestamp")
 #       cf2 <- CFtime(def, c)
 #       tp <- parse_timestamps(cf2, ts)
@@ -76,7 +76,7 @@ test_that("Gregorian/Julian calendar gap in the standard calendar", {
   t <- CFtime("days since 1582-10-01", "standard", 0:3)
   ts <- as_timestamp(t)
   expect_equal(ts[4], "1582-10-04")
-  expect_false(t$cal$POSIX_compatible(0:3))
+  expect_false(t$calendar$POSIX_compatible(0:3))
 
   t <- t + 4:10
   ts <- as_timestamp(t)
@@ -90,9 +90,9 @@ test_that("Gregorian/Julian calendar gap in the standard calendar", {
   expect_equal(ts[6], "1582-10-15")
 
   expect_false(t$POSIX_compatible())            # use t$offsets
-  expect_true(t$cal$POSIX_compatible(0:100))    # use supplied offsets
-  expect_true(t$cal$POSIX_compatible(-5:100))
-  expect_false(t$cal$POSIX_compatible(-6:100))
+  expect_true(t$calendar$POSIX_compatible(0:100))    # use supplied offsets
+  expect_true(t$calendar$POSIX_compatible(-5:100))
+  expect_false(t$calendar$POSIX_compatible(-6:100))
 })
 
 test_that("Leap seconds in the utc calendar work fine", {
@@ -133,7 +133,7 @@ test_that("Fractional time parts and replace Z timezone with offset", {
     for (u in 1:4) {
       def <- paste(CFt$units$name[u], "since 1978-08-20")
       t <- CFtime(def, c)
-      p <- t$cal$parse(offsets)
+      p <- t$calendar$parse(offsets)
       expect_equal(round(p$offset %% 1, 4), res[,u])
       expect_equal(p$tz, rep("+0000", 4))
     }
@@ -147,7 +147,7 @@ test_that("Fractional time parts and replace Z timezone with offset", {
   for (u in 1:4) {
     def <- paste(CFt$units$name[u], "since 1978-08-20")
     t <- CFtime(def, "utc")
-    p <- t$cal$parse(offsets)
+    p <- t$calendar$parse(offsets)
     expect_equal(round(p$offset %% 1, 4), res[,u])
   }