+++ NEWS.md
@@ -3,6 +3,7 @@
 Maintenance release
 
 * Support for CF-1.13.
+* Fields in `CFTime` and `CFClimatology` classes have been made private, corresponding active fields added.
 * `CFClimatology` instances can now be extended with the `+` operator.
 * When extending a `CFTime` or `CFClimatology` instance with another such instance using the `+` operator, bounds are preserved if both instances have bounds set.
 * Documentation updated.